outsourced deliveryNotes logic for usage in cronjob
