+++ plugins/Admin/resources/locales/default.pot
@@ -5,7 +5,7 @@
 msgid ""
 msgstr ""
 "Project-Id-Version: PROJECT VERSION\n"
-"POT-Creation-Date: 2022-01-29 09:39+0100\n"
+"POT-Creation-Date: 2022-01-29 10:14+0100\n"
 "PO-Revision-Date: YYYY-mm-DD HH:MM+ZZZZ\n"
 "Last-Translator: NAME <EMAIL@ADDRESS>\n"
 "Language-Team: LANGUAGE <EMAIL@ADDRESS>\n"
@@ -257,39 +257,6 @@ msgstr ""
 msgid "Edit_settings"
 msgstr ""
 
-msgid "Amount"
-msgstr ""
-
-msgid "Product"
-msgstr ""
-
-msgid "net_per_piece_abbr"
-msgstr ""
-
-msgid "Weight"
-msgstr ""
-
-msgid "Unit"
-msgstr ""
-
-msgid "Tax_rate"
-msgstr ""
-
-msgid "net"
-msgstr ""
-
-msgid "VAT"
-msgstr ""
-
-msgid "gross"
-msgstr ""
-
-msgid "Tax_rates_overview_table"
-msgstr ""
-
-msgid "Delivery_note"
-msgstr ""
-
 msgid "The_settings_of_manufacturer_{0}_have_been_changed."
 msgstr ""
 
@@ -342,6 +309,9 @@ msgid "The_reason_for_changing_the_member_is_mandatory."
 msgstr ""
 
 msgid "The_ordered_product_{0}_was_successfully_assigned_from_{1}_to_{2}."
+msgstr ""
+
+msgid "Amount"
 msgstr ""
 
 msgid "Reason"
@@ -818,6 +788,9 @@ msgstr ""
 msgid "ID"
 msgstr ""
 
+msgid "Weight"
+msgstr ""
+
 msgid "Associated_to_active_products?"
 msgstr ""
 
@@ -1283,6 +1256,9 @@ msgstr ""
 msgid "Net"
 msgstr ""
 
+msgid "VAT"
+msgstr ""
+
 msgid "Gross"
 msgstr ""
 
@@ -1598,13 +1574,22 @@ msgstr ""
 msgid "Test_order_list"
 msgstr ""
 
+msgid "Product"
+msgstr ""
+
 msgid "Test_invoice"
+msgstr ""
+
+msgid "Delivery_note"
 msgstr ""
 
 msgid "Manufacturer_profile"
 msgstr ""
 
 msgid "Copy_all_email_addresses"
+msgstr ""
+
+msgid "Tax_rate"
 msgstr ""
 
 msgid "Purchase_price_tax_excl"
@@ -1775,7 +1760,13 @@ msgstr ""
 msgid "Stock_product"
 msgstr ""
 
+msgid "gross"
+msgstr ""
+
 msgid "Surcharge"
+msgstr ""
+
+msgid "net"
 msgstr ""
 
 msgid "Selling_price_abbreviation"
