+++ plugins/Admin/resources/locales/en_US/admin.po
@@ -4,7 +4,7 @@
 msgid ""
 msgstr ""
 "Project-Id-Version: \n"
-"POT-Creation-Date: 2022-01-29 09:39+0100\n"
+"POT-Creation-Date: 2022-01-29 10:14+0100\n"
 "PO-Revision-Date: 2022-01-29 09:39+0100\n"
 "Last-Translator: \n"
 "Language-Team: LANGUAGE <EMAIL@ADDRESS>\n"
@@ -258,39 +258,6 @@ msgstr "Manufacturers"
 msgid "Edit_settings"
 msgstr "Edit settings"
 
-msgid "Amount"
-msgstr "Amount"
-
-msgid "Product"
-msgstr "Product"
-
-msgid "net_per_piece_abbr"
-msgstr "net/piece"
-
-msgid "Weight"
-msgstr "Weight"
-
-msgid "Unit"
-msgstr "Unit"
-
-msgid "Tax_rate"
-msgstr "Tax rate"
-
-msgid "net"
-msgstr "net"
-
-msgid "VAT"
-msgstr "VAT"
-
-msgid "gross"
-msgstr "gross"
-
-msgid "Tax_rates_overview_table"
-msgstr "Tax overview total"
-
-msgid "Delivery_note"
-msgstr "Delivery note"
-
 msgid "The_settings_of_manufacturer_{0}_have_been_changed."
 msgstr "The settings of manufacturer {0} have been changed."
 
@@ -344,6 +311,9 @@ msgstr "The reason for changing the member is mandatory."
 
 msgid "The_ordered_product_{0}_was_successfully_assigned_from_{1}_to_{2}."
 msgstr "The ordered product {0} was successfully assigned from {1} to {2}."
+
+msgid "Amount"
+msgstr "Amount"
 
 msgid "Reason"
 msgstr "Reason"
@@ -819,6 +789,9 @@ msgstr "Check and do not forget to click <b>Save</b> button."
 msgid "ID"
 msgstr "ID"
 
+msgid "Weight"
+msgstr "Weight"
+
 msgid "Associated_to_active_products?"
 msgstr "Associated with active products?"
 
@@ -1284,6 +1257,9 @@ msgstr "Invoice date"
 msgid "Net"
 msgstr "Net"
 
+msgid "VAT"
+msgstr "VAT"
+
 msgid "Gross"
 msgstr "Gross"
 
@@ -1599,14 +1575,23 @@ msgstr "Edit manufacturer settings"
 msgid "Test_order_list"
 msgstr "Test order list"
 
+msgid "Product"
+msgstr "Product"
+
 msgid "Test_invoice"
 msgstr "Test invoice"
+
+msgid "Delivery_note"
+msgstr "Delivery note"
 
 msgid "Manufacturer_profile"
 msgstr "Manufacturer profile"
 
 msgid "Copy_all_email_addresses"
 msgstr "Copy all e-mail adresses"
+
+msgid "Tax_rate"
+msgstr "Tax rate"
 
 msgid "Purchase_price_tax_excl"
 msgstr "Purchase price tax excl."
@@ -1776,8 +1761,14 @@ msgstr "Product declaration"
 msgid "Stock_product"
 msgstr "Stock product"
 
+msgid "gross"
+msgstr "gross"
+
 msgid "Surcharge"
 msgstr "Surcharge"
+
+msgid "net"
+msgstr "net"
 
 msgid "Selling_price_abbreviation"
 msgstr "SP"
@@ -2369,6 +2360,15 @@ msgstr "The delivered weight will eventually be adapted, so the price can change
 
 msgid "Product_ID"
 msgstr "ProductID"
+
+#~ msgid "net_per_piece_abbr"
+#~ msgstr "net/piece"
+
+#~ msgid "Unit"
+#~ msgstr "Unit"
+
+#~ msgid "Tax_rates_overview_table"
+#~ msgstr "Tax overview total"
 
 #, fuzzy
 #~ msgid "to_include_stock_products_in_order_lists."
